rename _op_item_type_registry to _item_type_registry
