+++ pyonepassword/op_items/item_types/api_credential.py
@@ -1,7 +1,7 @@
 import datetime
 
 from .._item_descriptor_registry import op_register_item_descriptor_type
-from .._op_item_type_registry import op_register_item_type
+from .._item_type_registry import op_register_item_type
 from ._item_base import OPAbstractItem
 from ._item_descriptor_base import OPAbstractItemDescriptor
