+++ pyonepassword/op_items/item_types/credit_card.py
@@ -1,5 +1,5 @@
 from .._item_descriptor_registry import op_register_item_descriptor_type
-from .._op_item_type_registry import op_register_item_type
+from .._item_type_registry import op_register_item_type
 from ..fields_sections.item_field_base import OPItemField
 from ..fields_sections.item_section import OPSection
 from ._item_base import OPAbstractItem
