+++ tests/test_non_conformant_data/test_duplicate_sections.py
@@ -28,7 +28,7 @@ from pyonepassword.api.validation import (
     get_relaxed_validation_for_class,
     set_relaxed_validation_for_class
 )
-from pyonepassword.op_items._op_item_type_registry import OPItemFactory
+from pyonepassword.op_items._item_type_registry import OPItemFactory
 from pyonepassword.op_items.item_validation_policy import (
     _OPItemValidationPolicy
 )
